#84: Implement turning at arbitrary angles. Add new lgo files to test these. We can now make a spirograph and other pretty pictures. Handle case where incorrect CLI argument is passed for the factory. TDD: Current tests are passing. Adapt Raylib movmement and render test to pass.
Driver: ag20751
Navigator: mw20619, sg21642


git-svn-id: https://csee.essex.ac.uk/svn/ce320-06/trunk@111 48541511-ed4e-4a0e-8099-d0968ce73c23

+++ src/factory/pd_factory.h
@@ -34,10 +34,10 @@ class PDFactory : public UIFactory {
   }
 
   std::unique_ptr<Movement> createMovement(
-      int y_location, int x_location, int y_safe_zone, int x_safe_zone,
+      float y_location, float x_location, int y_safe_zone, int x_safe_zone,
       CharacterAssembler* assembler) override {
     return std::make_unique<PDMovement>(
-        y_location, x_location, y_safe_zone, x_safe_zone,
+        (int)y_location, (int)x_location, y_safe_zone, x_safe_zone,
         static_cast<PDCharacterAssembler*>(assembler));
   }
 };
