+++ src/movement/movement.h
@@ -10,8 +10,8 @@ class Movement {
   virtual void TurnRight90() = 0;
   virtual void TurnLeft90() = 0;
 
-  virtual int GetXLocation() const = 0;
-  virtual int GetYLocation() const = 0;
+  virtual float GetXLocation() const = 0;
+  virtual float GetYLocation() const = 0;
 
   virtual void SetPenPosition(PenPosition pen_position) = 0;
   virtual PenPosition GetPenPosition() const = 0;
